+++ src/chat/functions/index.ts
@@ -25,6 +25,7 @@ export { generateMessageID } from './generateMessageID';
 export { get } from './get';
 export { getMessageById } from './getMessageById';
 export { getMessages, GetMessagesOptions } from './getMessages';
+export { ChatListOptions, list } from './list';
 export { markIsComposing } from './markIsComposing';
 export { markIsPaused } from './markIsPaused';
 export { markIsRead } from './markIsRead';
@@ -41,6 +42,10 @@ export {
   prepareMessageButtons,
 } from './prepareMessageButtons';
 export { prepareRawMessage } from './prepareRawMessage';
+export {
+  PoolMessageOptions,
+  sendCreatePollMessage,
+} from './sendCreatePollMessage';
 export {
   AudioMessageOptions,
   AutoDetectMessageOptions,